+++ snapshot-tests-html/src/main/java/de/skuzzle/test/snapshots/data/html/HtmlStructuralAssertions.java
@@ -1,14 +1,29 @@
 package de.skuzzle.test.snapshots.data.html;
 
+import java.util.function.Consumer;
+
 import org.jsoup.Jsoup;
 import org.jsoup.helper.W3CDom;
 import org.jsoup.nodes.Document;
+import org.xmlunit.assertj.CompareAssert;
 import org.xmlunit.assertj.XmlAssert;
+import org.xmlunit.diff.DifferenceEvaluator;
 
 import de.skuzzle.test.snapshots.SnapshotException;
 import de.skuzzle.test.snapshots.StructuralAssertions;
+import de.skuzzle.test.snapshots.validation.Arguments;
 
 final class HtmlStructuralAssertions implements StructuralAssertions {
+
+    private final DifferenceEvaluator differenceEvaluator;
+    private final Consumer<CompareAssert> compareAssertConsumer;
+
+    public HtmlStructuralAssertions(Consumer<CompareAssert> compareAssertConsumer,
+            DifferenceEvaluator differenceEvaluator) {
+        this.differenceEvaluator = differenceEvaluator;
+        this.compareAssertConsumer = Arguments.requireNonNull(compareAssertConsumer,
+                "compareAssertConsumer must not be null");
+    }
 
     @Override
     public void assertEquals(String storedSnapshot, String serializedActual) throws AssertionError, SnapshotException {
@@ -16,9 +31,14 @@ final class HtmlStructuralAssertions implements StructuralAssertions {
         final Document parsedSnapshot = Jsoup.parse(storedSnapshot);
         final Document parsedActual = Jsoup.parse(serializedActual);
 
-        XmlAssert.assertThat(w3cDom.fromJsoup(parsedSnapshot))
-                .and(w3cDom.fromJsoup(parsedActual))
-                .areIdentical();
+        CompareAssert compareAssert = XmlAssert
+                .assertThat(w3cDom.fromJsoup(parsedSnapshot))
+                .and(w3cDom.fromJsoup(parsedActual));
+        if (differenceEvaluator != null) {
+            compareAssert = compareAssert.withDifferenceEvaluator(differenceEvaluator);
+        }
+
+        compareAssertConsumer.accept(compareAssert);
     }
 
 }